kz - patches for loading java in regionalGroundMotion

+++ modules/performRegionalEventSimulation/regionalGroundMotion/HazardSimulation.py
@@ -107,23 +107,21 @@ if __name__ == '__main__':
             subprocess.check_call([sys.executable, "-m", "pip", "install", p])
 
     # set up environment
-    if opensha_flag or opensha_vs30_flag:
-        if importlib.util.find_spec('jpype') is None:
-            subprocess.check_call([sys.executable, "-m", "pip", "install", "JPype1"])
-        import jpype
-        from jpype import imports
-        from jpype.types import *
-        jpype.addClassPath('./lib/OpenSHA-1.5.2.jar')
-        jpype.startJVM("-Xmx8G", convertStrings=False)
+    if importlib.util.find_spec('jpype') is None:
+        subprocess.check_call([sys.executable, "-m", "pip", "install", "JPype1"])
+    import jpype
+    from jpype import imports
+    from jpype.types import *
+    jpype.addClassPath('./lib/OpenSHA-1.5.2.jar')
+    jpype.startJVM("-Xmx8G", convertStrings=False)
     if oq_flag:
         # data dir
-        os.environ['OQ_DATADIR'] = os.path.join(os.path.abspath(os.path.dirname(output_dir)), 'oqdata')
+        os.environ['OQ_DATADIR'] = os.path.join(os.path.abspath(os.path.dirname(input_dir)), 'oqdata')
         print('HazardSimulation: local OQ_DATADIR = '+os.environ.get('OQ_DATADIR'))
-        try:
-            os.makedirs(os.environ.get('OQ_DATADIR'))
-        except:
+        if os.path.exists(os.environ.get('OQ_DATADIR')):
             print('HazardSimulation: local OQ folder already exists, overwiting it now...')
             shutil.rmtree(os.environ.get('OQ_DATADIR'))
+        os.makedirs(os.environ.get('OQ_DATADIR'))    
     
     # import modules
     from CreateStation import *
